+++ src/Web/Grand.Web.Admin/Services/KnowledgebaseViewModelService.cs
@@ -1,10 +1,10 @@
 ﻿using Grand.Business.Core.Extensions;
+using Grand.Business.Core.Interfaces.Cms;
 using Grand.Business.Core.Interfaces.Common.Directory;
 using Grand.Business.Core.Interfaces.Common.Localization;
 using Grand.Business.Core.Interfaces.Common.Logging;
 using Grand.Business.Core.Interfaces.Common.Seo;
 using Grand.Business.Core.Interfaces.Customers;
-using Grand.Business.Core.Interfaces.Marketing.Knowledgebase;
 using Grand.Domain.Knowledgebase;
 using Grand.Domain.Seo;
 using Grand.Infrastructure;
