+++ src/Web/Grand.Web/Extensions/KnowledgebaseExtensions.cs
@@ -1,6 +1,6 @@
 ﻿using Grand.Business.Core.Extensions;
+using Grand.Business.Core.Interfaces.Cms;
 using Grand.Business.Core.Interfaces.Common.Security;
-using Grand.Business.Core.Interfaces.Marketing.Knowledgebase;
 using Grand.Domain.Knowledgebase;
 using Grand.Infrastructure;
