+++ src/Web/Grand.Web/Features/Handlers/Common/GetSitemapHandler.cs
@@ -2,7 +2,6 @@
 using Grand.Business.Core.Interfaces.Catalog.Products;
 using Grand.Business.Core.Interfaces.Cms;
 using Grand.Business.Core.Extensions;
-using Grand.Business.Core.Interfaces.Marketing.Knowledgebase;
 using Grand.Infrastructure.Caching;
 using Grand.Domain.Blogs;
 using Grand.Domain.Common;
